requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.24.3
torch==2.1.1
scikit-learn==1.3.2
redis==5.0.1
asyncpg==0.29.0
pydantic==2.5.0
python-multipart==0.0.6
PyJWT==2.8.0
httpx==0.25.2
websockets==12.0
prometheus-client==0.19.0
motor==3.1.2
pymongo==4.3.3
